+++ release_notes.md
@@ -1,10 +1,17 @@
 # Release Notes
 
+## Version 0.0.242 (November 14, 2025)
+- Enahnced markdown parser
+- add support for Dataform and SQLX parsing for context enrichment
+- Added GPT-5.1
+- Improved metadata indexing service
+- Improved UI for reports and tables
+
 ## Version 0.0.241 (November 14, 2025)
 - Optimize datbase migrations to include report_type
 - Wrap maintenance job with guardrails
 
-## Version 0.0.240 (November 14, 2025)
+## Version 0.0.240 (November 13, 2025)
 - Introducing Evals! You can now create and run custom sets of tests on demand to assess system performance. Define your own test cases and assertions, such as:
   - User prompts triggering create_data on table1 and table2
   - Validating that specific data columns (e.g., a, b, c) are present